requirements.txt
ExtractTable>=1.2.0
camelot-py>=0.7.3
